_legacy/requirement.txt
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
markdown-it-py==4.0.0
mdurl==0.1.2
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
Pygments==2.19.2
PyMuPDF==1.26.6
pypdfium2==5.1.0
PyYAML==6.0.3
fastapi==0.111.0
uvicorn==0.30.1
rich==14.2.0
shellingham==1.5.4
typer==0.20.0
typing_extensions==4.15.0
pytesseract==0.3.10
pytest==8.3.2
